Pacman AI works! Try it out for yourself!
Only two decent levels are currently in the game. The rest are test
levels. No scoring system yet, but that will come soon (probably). Let
me know if you beat the second level, cuz it was too hard for me to
beat. Press period and comma to cycle between the levels if you want to
go to a specific one. Have fun!